Feat: user now has their own rating, post, edit and delete comment in backend.

+++ ecommerce-server-app/app/Http/Controllers/ProductReviewController.php
@@ -4,20 +4,33 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\ProductReview;
+use App\Models\User;
 
 class ProductReviewController extends Controller
 {
     public function store(Request $request, $id)
     {
+
         $validatedData = $request->validate([
             'rating' => 'required|integer',
-            'reviewText' => 'required|string',
+            'review_text' => 'required|string',
         ]);
+
+        $userId = $request->input('userId');
+
+        $user = User::find($userId);
+
+        if (!$user) {
+            return response()->json(['error' => 'User not found'], 404);
+        }
 
         $productReview = new ProductReview;
         $productReview->rating = $validatedData['rating'];
-        $productReview->reviewText = $validatedData['reviewText'];
+        $productReview->review_text = $validatedData['review_text'];
+        $productReview->user_id = $userId;
         $productReview->product_id = $id;
+
+
 
         if ($productReview->save()) {
             return response()->json(['message' => 'Review has been submitted successfully', 'commentId' => $productReview->commentId], 201);
@@ -26,16 +39,33 @@ class ProductReviewController extends Controller
         }
     }
 
-    public function showUserComment($id)
+    public function showUserComment($id, Request $request)
     {
-        $commentId = ProductReview::where('product_id', $id)->pluck('commentId');
+        $userId = $request->input('userId');
+
+        $user = User::find($userId);
+
+        if (!$user) {
+            return response()->json(['error' => 'User not found'], 404);
+        }
+
+
+        $productReviews = ProductReview::where('product_id', $id)
+            ->where('user_id', $userId)
+            ->get();
+
+        if ($productReviews->isEmpty()) {
+            return response()->json(['error' => 'No reviews found for this product by the user'], 404);
+        }
+
+        $commentId = ProductReview::where('product_id', $id)->pluck('comment_id');
         $rating = ProductReview::where('product_id', $id)->avg('rating');
-        $reviewText = ProductReview::where('product_id', $id)->pluck('reviewText')->toArray();
-    
+        $reviewText = ProductReview::where('product_id', $id)->pluck('review_text')->toArray();
+
         return response()->json([
-            'commentId' => $commentId, 
+            'comment_id' => $commentId,
             'rating' => $rating,
-            'reviewText' => $reviewText,
+            'reviewT_text' => $reviewText,
         ], 200);
     }
 
@@ -43,7 +73,7 @@ class ProductReviewController extends Controller
     {
         $validatedData = $request->validate([
             'rating' => 'required|integer',
-            'reviewText' => 'required|string',
+            'review_text' => 'required|string',
         ]);
 
         $productReview = ProductReview::find($id);
@@ -52,8 +82,25 @@ class ProductReviewController extends Controller
             return response()->json(['message' => 'Review not found'], 404);
         }
 
+        $userId = $request->input('userId');
+
+        $user = User::find($userId);
+
+        if (!$user) {
+            return response()->json(['error' => 'User not found'], 404);
+        }
+
+
+        $productReviews = ProductReview::where('comment_id', $id)
+            ->where('user_id', $userId)
+            ->get();
+
+        if ($productReviews->isEmpty()) {
+            return response()->json(['error' => 'No reviews found for this product by the user'], 404);
+        }
+
         $productReview->rating = $validatedData['rating'];
-        $productReview->reviewText = $validatedData['reviewText'];
+        $productReview->review_text = $validatedData['review_text'];
 
         if ($productReview->save()) {
             return response()->json(['message' => 'Review has been updated successfully'], 200);
@@ -62,10 +109,25 @@ class ProductReviewController extends Controller
         }
     }
 
-    public function deleteComment($id)
+    public function deleteComment(Request $request, $id)
     {
         $productReview = ProductReview::find($id);
+        $userId = $request->input('userId');
 
+        $user = User::find($userId);
+
+        if (!$user) {
+            return response()->json(['error' => 'User not found'], 404);
+        }
+
+
+        $productReviews = ProductReview::where('comment_id', $id)
+            ->where('user_id', $userId)
+            ->get();
+
+        if ($productReviews->isEmpty()) {
+            return response()->json(['error' => 'No reviews found for this product by the user'], 404);
+        }
         if ($productReview) {
             $productReview->delete();
             return response()->json(['message' => 'Comment deleted successfully'], 200);
